refactor: migrate flat root screens to (stack) group folder structure
- Move 9 root-level screens (create, write, edit, editWithAi, generate,
  confirmPhoto, search, loading, loading/) into app/(stack)/ group folder
- Move loading/ subfolder into (stack)/loading/ with _layout.js
  (gestureEnabled: false applied via layout instead of per-screen)
- Convert loading.js → loading/index.js to resolve naming conflict
  with loading/ directory
- Fix diary/_layout.js: remove orphaned Stack.Screen without parent Stack
- Update all relative imports after path depth changes
- Replace /home redirects with /(tabs) and delete home.js (was a no-op redirect)
- Fix login.js: /terms → /(onboarding)/terms
