+++ app/(stack)/loading/_layout.js
@@ -1,0 +1,12 @@
+import { Stack } from "expo-router";
+
+export default function LoadingLayout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+        gestureEnabled: false
+      }}
+    />
+  );
+}
