+++ app/diary/_layout.js
@@ -1,5 +1,4 @@
 // app/diary/_layout.js
-import React from "react";
 import { Stack } from "expo-router";
 
 export default function DiaryLayout() {
@@ -8,7 +7,8 @@ export default function DiaryLayout() {
       <Stack
         screenOptions={{
           headerShown: false,
-          contentStyle: { backgroundColor: "#FCF9F4" }
+          contentStyle: { backgroundColor: "#FCF9F4" },
+          gestureEnabled: false
         }}
       />
     </>
